Add instance index to log directory/archive names
This add the instance index, such as win-3, to the log directory
names and zip file names. With this information, it easier to match
up Terraforms reporting with the related S3 artifacts. For example,
Terraform may report an error related to instance win.1 or win[1].
Now the S3 log directory for that instance and the archive file will
include win-1.